refacto SecurityPlatformAnalysis
