+++ opencti-platform/opencti-front/src/private/components/entities/securityPlatforms/SecurityPlatform.tsx
@@ -60,7 +60,7 @@ export const securityPlatformFragment = graphql`
       }
     }
     workflowEnabled
-...SecurityPlatformDetails_securityPlatform
+    ...SecurityPlatformDetails_securityPlatform
   }
 `;
