+++ opencti-platform/opencti-front/src/private/components/entities/securityPlatforms/SecurityPlatformAnalysis.tsx
@@ -1,0 +1,32 @@
+import React, { FunctionComponent } from 'react';
+import { graphql, createFragmentContainer } from 'react-relay';
+import { SecurityPlatformDetails_securityPlatform$data } from '@components/entities/securityPlatforms/__generated__/SecurityPlatformDetails_securityPlatform.graphql';
+import StixCoreObjectOrStixCoreRelationshipContainers from '../../common/containers/StixCoreObjectOrStixCoreRelationshipContainers';
+
+interface SecurityPlatformAnalysisComponentProps {
+  securityPlatform: SecurityPlatformDetails_securityPlatform$data;
+}
+
+const SecurityPlatformAnalysisComponent: FunctionComponent<SecurityPlatformAnalysisComponentProps> = ({ securityPlatform }) => {
+  return (
+    <StixCoreObjectOrStixCoreRelationshipContainers
+      stixDomainObjectOrStixCoreRelationship={securityPlatform}
+    />
+  );
+};
+
+const SecurityPlatformAnalysis = createFragmentContainer(
+  SecurityPlatformAnalysisComponent,
+  {
+    securityPlatform: graphql`
+      fragment SecurityPlatformAnalysis_securityPlatform on SecurityPlatform {
+        id
+        name
+        x_opencti_aliases
+        x_opencti_graph_data
+      }
+    `,
+  },
+);
+
+export default SecurityPlatformAnalysis;
